requirements.txt
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.40.0
diffusers>=0.27.0
accelerate>=0.28.0
numpy>=1.26.0
scipy>=1.13.0
Pillow>=10.0.0
opencv-python>=4.9.0
einops>=0.7.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
websockets>=12.0
tqdm>=4.66.0
matplotlib>=3.8.0
scikit-learn>=1.4.0
msgpack>=1.0.8
pytest>=8.1.0
imageio>=2.34.0
